!!![TASK] Set default value for pages.tx_in2responsible_check to "not checked"
Before TYPO3 13 in2responsible used an ext_tables.sql with default value 0 and a TCA configuration with default value 1 for pages.tx_in2responsible_check. So, when in2responsible was installed in a large system, a new field was added to pages but with value 0. For new pages 1 was set by default.
Now for TYPO3 13, we want to get rid of ext_tables.sql. This means we have to decide if we want a 0 or a 1 for the field in database and the same value for new page records.
And we think it's more helpful to set this value per default to 0.
If someone wants to overrule this, it can be done with a sitepacke by manipulating $GLOBALS['TCA']